+++ doc/getUser.md
@@ -1,50 +1,36 @@
-# Retriving user details
-
-Returns a JSON with information about a single user
-
+# Creating user 
+Returns a URL containing the ID of the new created user 
 * **URL:** <br>
-    /users/:id
-
+    /users/
 * **Method:** <br>
-    `GET`
-
+    `POST`
 * **URL Params** <br>
+	None
 **Required:** <br>
-    `id = [integer]`
-
+    `Data in body`
 * **Data Params** <br>
-    None
-
+    All the user details completed in the input fields given using body
 * **Success Response:** <br>
     * **Code:** 200 <br>
       **Content:** 
-    ```javascript
+      ```javascript
         {
             userId: 1, 
-            first_name: "Catalin",
-            last_name: "Popescu",
-            email: "catalin.popescu@gmail.com",
-            phone_number: "0745231445"
-    }
-    ```
-
+            first_name: “Ion”,
+            last_name: “Neculce”,
+            email: "cronicar@gmail.com",
+            phone_number: "0720557248”
+       }
+       ```
 * **Error Response:** <br>
-    * **Code:** 404 NOT FOUND <br>
+    * **Code:** 400 FAILED CLIENT REQUEST <br>
       **Content:** 
-    ```javascript
-    {
-        error: "User not found"
-    }
-    ```
-    * **Code:** 500 INTERNAL SERVER ERROR <br>
-      **Content:** 
-    ```javascript
-    {
-        error: "Internal server error"
-    }
-    ```
- 
-* **Sample Call:** <br>
-    ```javascript
-    fetch("localhost:3000/users/1");
-    ```
+      ```javascript
+      {
+         error_fname: “Ati introdus prenumele gresit”,
+	    error_lname: “Ati introdus numele gresit”,
+	    error_email: “Ati introdus emailul gresit”,
+	    error_phone: “Ati introdus gresit numarul de telefon”
+      }
+      ```
+
