+++ src/actions/retrieveUser.js
@@ -1,0 +1,4 @@
+let retrieveUser = () =>{
+
+};
+module.exports = retrieveUser;
