+++ src/main.js
@@ -1,12 +1,10 @@
 let express = require("express");
 let bodyParser = require("body-parser")
 let usersRouter = require("./routes/users");
-let daysRouter = require("./routes/days");
 let exampleRouter = require("./routes/example");
 
 let app = express();
 app.use(bodyParser.json());
-app.use("/days", daysRouter);
 app.use("/users", usersRouter);
 app.use("/api", exampleRouter);
