--- src/routes/days.js
@@ -1,3 +1,0 @@
-let router = require("express").Router();
-
-module.exports = router;
